Add more information to annotation errors Add lots of new StorageSettings store and delete calls now return a CompletableFuture that is completed when the operation finishes YAMLContainer now supports lists of DecomposedObjects Fix issue with JSONContainer returning null when retrieving a list handler Revise the caching/flushing systems, and how often the complete is called on containers.

+++ storage-api/src/main/java/me/dessie/dessielib/storageapi/storage/container/StorageContainer.java
@@ -2,9 +2,11 @@ package me.dessie.dessielib.storageapi.storage.container;
 
 import me.dessie.dessielib.storageapi.StorageAPI;
 import me.dessie.dessielib.storageapi.storage.cache.CachedObject;
+import me.dessie.dessielib.storageapi.storage.cache.FlushTask;
 import me.dessie.dessielib.storageapi.storage.cache.StorageCache;
 import me.dessie.dessielib.storageapi.storage.container.hooks.DeleteHook;
 import me.dessie.dessielib.storageapi.storage.container.hooks.RetrieveHook;
+import me.dessie.dessielib.storageapi.storage.container.hooks.StorageHook;
 import me.dessie.dessielib.storageapi.storage.container.hooks.StoreHook;
 import me.dessie.dessielib.storageapi.storage.decomposition.DecomposedObject;
 import me.dessie.dessielib.storageapi.storage.decomposition.StorageDecomposer;
@@ -12,8 +14,8 @@ import me.dessie.dessielib.storageapi.storage.settings.StorageSettings;
 import org.bukkit.Bukkit;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
@@ -30,9 +32,9 @@ public abstract class StorageContainer {
     /**
      * Creates a StorageContainer with a default {@link StorageSettings}.
      *
-     * @see StorageSettings#StorageSettings() for the default settings.
+     * @see StorageSettings#StorageSettings()
      */
-    public StorageContainer() {
+    protected StorageContainer() {
         this(new StorageSettings());
     }
 
@@ -41,7 +43,7 @@ public abstract class StorageContainer {
      *
      * @param settings The StorageSettings for the container.
      */
-    public StorageContainer(StorageSettings settings) {
+    protected StorageContainer(StorageSettings settings) {
         if(!StorageAPI.isRegistered()) {
             throw new NullPointerException("You need to register your plugin before creating a StorageContainer!");
         }
@@ -49,15 +51,7 @@ public abstract class StorageContainer {
         Objects.requireNonNull(settings, "Settings cannot be null!");
 
         this.settings = settings;
-        this.cache = new StorageCache(this.getSettings().getCacheDuration());
-
-        //The repeating task to push add modified changes to the data source.
-        //Uses the settings for the length.
-        if(this.getSettings().getUpdate() > 0) {
-            Bukkit.getScheduler().runTaskTimer(StorageAPI.getPlugin(), () -> {
-                this.getCache().flush(this);
-            }, this.getSettings().getUpdate() * 20L, this.getSettings().getUpdate() * 20L);
-        }
+        this.cache = new StorageCache(this, this.getSettings().getCacheDuration());
     }
 
     /**
@@ -127,7 +121,7 @@ public abstract class StorageContainer {
      * Sets data into a cache that will eventually be updated into the data structure.
      * Data that is set may be lost in the case of a hard crash.
      *
-     * To determine how often this data is pushed to the structure, change the {@link StorageSettings}.
+     * To determine how often this data is pushed to the structure, use {@link StorageSettings#setFlushRate(int)}
      *
      * @see StorageContainer#store(String, Object) for storing data within the source.
      *
@@ -141,19 +135,33 @@ public abstract class StorageContainer {
             throw new IllegalArgumentException(data.getClass() + " is not a supported storage class. Create a StorageDecomposer to implement behavior!");
         }
 
-        this.cache(path, data);
+        this.cacheStore(path, data);
 
         this.getCache().getSetCache().put(path, data);
         this.getCache().getRemoveCache().remove(path);
     }
 
     /**
-     * Puts data into a cache that will eventually be updated into the data structure to remove them.
+     * Sets data into a cache that will eventually be updated into the data structure.
+     * Data that is set may be lost in the case of a hard crash.
+     *
+     * To determine how often this data is pushed to the structure, use {@link StorageSettings#setFlushRate(int)}
+     *
+     * @see StorageContainer#storeAll(Map) for storing data within the source.
+     *
+     * @param data A map with all the paths and objects to set.
+     */
+    public void setAll(Map<String, Object> data) {
+        data.forEach(this::set);
+    }
+
+    /**
+     * Puts a path into a cache that will eventually be updated into the data structure to remove them.
      * Data that is removed may be lost in the case of a hard crash.
      *
-     * To determine how often this data is pushed to the structure, change the {@link StorageSettings}.
+     * To determine how often this data is pushed to the structure, use {@link StorageSettings#setFlushRate(int)}
      *
-     * @see StorageContainer#delete(String)  for immediately deleting data within the source.
+     * @see StorageContainer#delete(String) for immediately deleting data within the source.
      *
      * @param path The path of the data to remove.
      */
@@ -167,61 +175,78 @@ public abstract class StorageContainer {
     }
 
     /**
+     * Puts paths into a cache that will eventually be updated into the data structure to remove them.
+     * Data that is removed may be lost in the case of a hard crash.
+     *
+     * To determine how often this data is pushed to the structure, use {@link StorageSettings#setFlushRate(int)}
+     *
+     * @see StorageContainer#deleteAll(List)  for immediately deleting data within the source.
+     *
+     * @param paths The path of the data to remove.
+     */
+    public void removeAll(List<String> paths) {
+        paths.forEach(this::remove);
+    }
+
+    /**
      * Stores data to the data structure. This method is executed asynchronously.
      *
      * @see StorageContainer#set(String, Object) for caching objects instead of writing directly to the structure.
      *
      * @param path The path to store the data to.
      * @param data The data to store in the file format.
+     *
+     * @return A {@link CompletableFuture} that will be completed once the async storage has been finished.
+     *         This will complete before {@link StorageHook#complete()} has been called.
      */
-    public void store(String path, Object data) {
-        Objects.requireNonNull(path, "Cannot store to null path!");
+    public CompletableFuture<Void> store(String path, Object data) {
+        CompletableFuture<Void> future = this.storeData(path, data);
 
-        if(!isSupported(data.getClass())) {
-            throw new IllegalArgumentException(data.getClass() + " is not a supported storage class. Create a StorageDecomposer to implement behavior!");
-        }
+        if(this.getCache().getFlushTask().canFlush()) {
+            future.thenRunAsync(() -> this.storeHook().complete());
+            this.getCache().getFlushTask().resetFlushCooldown();
 
-        StorageDecomposer<?> decomposer = getDecomposer(data.getClass());
-        DecomposedObject object = null;
-
-        //Cache the data.
-        if(decomposer != null) {
-            String decomposePath = path + ".%path%";
-            object = decomposer.applyDecompose(data);
-            for(String decomposedPath : object.getDecomposedMap().keySet()) {
-                this.cache(decomposePath.replace("%path%", decomposedPath), object.getDecomposedMap().get(decomposedPath));
-            }
+            return future;
         } else {
-            this.cache(path, data);
+            CompletableFuture<Void> waiting = new CompletableFuture<>();
+            this.getCache().getFlushTask().addFuture(waiting);
+
+            //Wait for both the waiting task and the actual storage task to complete.
+            return CompletableFuture.allOf(waiting, future);
         }
 
-        DecomposedObject finalObject = object;
-        Bukkit.getScheduler().runTaskAsynchronously(StorageAPI.getPlugin(), () -> {
-            if(decomposer != null) {
-                //Add the placeholder for each decomposed path.
-                String decomposePath = path + ".%path%";
-                for(String decomposedPath : finalObject.getDecomposedMap().keySet()) {
-                    String compiledPath = decomposePath.replace("%path%", decomposedPath);
-                    Object decomposedObject = finalObject.getDecomposedMap().get(decomposedPath);
+    }
 
-                    if(decomposedObject != null && StorageContainer.getDecomposer(decomposedObject.getClass()) != null) {
-                        this.store(compiledPath, decomposedObject);
-                        continue;
-                    }
+    /**
+     * Stores data to the data structure. This method is executed asynchronously.
+     *
+     * @see StorageContainer#setAll(Map) for caching objects instead of writing directly to the structure.
+     *
+     * @param data A map with all the paths and objects to store.
+     * @return A {@link CompletableFuture} that will be completed once all data has been stored.
+     *         This will complete before {@link StorageHook#complete()} has been called.
+     */
+    public CompletableFuture<Void> storeAll(Map<String, Object> data) {
+        List<CompletableFuture<Void>> futures = new ArrayList<>();
 
-                    this.storeHook().getConsumer().accept(compiledPath, decomposedObject);
-                }
-            } else if (this instanceof ArrayContainer arrayContainer && arrayContainer.isList(data)) {
-                this.storeHook().getConsumer().accept(path, arrayContainer.handleList(data));
-            } else {
-                this.storeHook().getConsumer().accept(path, data);
-            }
-            this.storeHook().complete();
-        });
+        for(Map.Entry<String, Object> entry : data.entrySet()) {
+            futures.add(this.storeData(entry.getKey(), entry.getValue()));
+        }
 
-        //No need to update these later, since this should overwrite them.
-        this.getCache().getSetCache().remove(path);
-        this.getCache().getRemoveCache().remove(path);
+        CompletableFuture<Void> future = CompletableFuture.allOf(futures.toArray(new CompletableFuture[] {}));
+
+        if(this.getCache().getFlushTask().canFlush()) {
+            future.thenRunAsync(() -> this.storeHook().complete());
+            this.getCache().getFlushTask().resetFlushCooldown();
+
+            return future;
+        } else {
+            CompletableFuture<Void> waiting = new CompletableFuture<>();
+            this.getCache().getFlushTask().addFuture(waiting);
+
+            //Wait for both the waiting task and the actual storage task to complete.
+            return CompletableFuture.allOf(waiting, future);
+        }
     }
 
     /**
@@ -230,18 +255,72 @@ public abstract class StorageContainer {
      * @see StorageContainer#remove(String) for removing data using a cache. (Recommended)
      *
      * @param path The path to remove.
+     * @return A {@link CompletableFuture} that will be completed once the async deletion has been finished.
+     *         This will complete before {@link StorageHook#complete()} has been called.
      */
-    public void delete(String path) {
+    public CompletableFuture<Void> delete(String path) {
         Objects.requireNonNull(path, "Cannot delete from null path!");
+
+        CompletableFuture<Void> future = new CompletableFuture<>();
 
         Bukkit.getScheduler().runTaskAsynchronously(StorageAPI.getPlugin(), () -> {
             this.deleteHook().getConsumer().accept(path);
-            this.deleteHook().complete();
+
+            if(this.getCache().getFlushTask().canFlush()) {
+                this.deleteHook().complete();
+                this.getCache().getFlushTask().resetFlushCooldown();
+
+                future.complete(null);
+            } else {
+                CompletableFuture<Void> waiting = new CompletableFuture<>();
+                this.getCache().getFlushTask().addFuture(waiting);
+
+                waiting.thenRun(() -> future.complete(null));
+            }
         });
 
         //This should overwrite anything we've already cached to do.
         this.getCache().getRemoveCache().remove(path);
         this.getCache().getSetCache().remove(path);
+
+        return future;
+    }
+
+    /**
+     * Removes multiple paths from the data source. This method is executed asynchronously.
+     * Useful if you're removing lots of items and don't want to call {@link DeleteHook#complete()} constantly.
+     *
+     * @see StorageContainer#removeAll(List) for removing data using a cache. (Recommended)
+     *
+     * @param paths The paths to remove.
+     * @return A {@link CompletableFuture} that will be completed once the async deletion has been finished.
+     */
+    public CompletableFuture<Void> deleteAll(List<String> paths) {
+        Objects.requireNonNull(paths, "Cannot delete from null path!");
+
+        CompletableFuture<Void> future = new CompletableFuture<>();
+
+        Bukkit.getScheduler().runTaskAsynchronously(StorageAPI.getPlugin(), () -> {
+            paths.forEach(p -> this.deleteHook().getConsumer().accept(p));
+
+            if(this.getCache().getFlushTask().canFlush()) {
+                this.deleteHook().complete();
+                this.getCache().getFlushTask().resetFlushCooldown();
+
+                future.complete(null);
+            } else {
+                CompletableFuture<Void> waiting = new CompletableFuture<>();
+                this.getCache().getFlushTask().addFuture(waiting);
+
+                waiting.thenRun(() -> future.complete(null));
+            }
+        });
+
+        //This should overwrite anything we've already cached to do.
+        this.getCache().getRemoveCache().removeAll(paths);
+        this.getCache().getSetCache().keySet().removeIf(paths::contains);
+
+        return future;
     }
 
     /**
@@ -268,7 +347,7 @@ public abstract class StorageContainer {
 
         T obj = (T) this.retrieveHook().getFunction().apply(path);
         this.retrieveHook().complete();
-        this.cache(path, obj);
+        this.cacheRetrieve(path, obj);
         return obj;
     }
 
@@ -451,7 +530,34 @@ public abstract class StorageContainer {
      */
     public void cache(String path, Object object) {
         Objects.requireNonNull(path, "Cannot cache null path!");
+
+        if(!this.getSettings().isUseCache()) return;
+
         this.getCache().cache(path, object);
+    }
+
+    /**
+     * Attempts to cache an object, but obeys the {@link StorageSettings#isCachedOnRetrieve()} setting.
+     *
+     * @param path The path to cache to
+     * @param object The object to cache
+     */
+    public void cacheRetrieve(String path, Object object) {
+        if(this.getSettings().isCachedOnRetrieve()) {
+            this.cache(path, object);
+        }
+    }
+
+    /**
+     * Attempts to cache an object, but obeys the {@link StorageSettings#isCachedOnStore()} setting.
+     *
+     * @param path The path to cache to
+     * @param object The object to cache
+     */
+    public void cacheStore(String path, Object object) {
+        if(this.getSettings().isCachedOnStore()) {
+            this.cache(path, object);
+        }
     }
 
     /**
@@ -491,7 +597,15 @@ public abstract class StorageContainer {
      * @see StorageContainer#remove(String) for adding paths into the Remove cache.
      */
     public void flush() {
-        this.getCache().flush(this);
+        StorageCache cache = this.getCache();
+        FlushTask task = cache.getFlushTask();
+
+        if(task.canFlush()) {
+            cache.flush();
+            task.resetFlushCooldown();
+        } else {
+            task.queueFlush();
+        }
     }
 
     /**
@@ -515,7 +629,6 @@ public abstract class StorageContainer {
         if(clazz == String.class) return true;
         if(getDecomposer(clazz) != null) return true;
         if(StorageAPI.getWrappers().containsValue(clazz)) return true;
-        if(this instanceof ArrayContainer && (clazz.isArray() || Collection.class.isAssignableFrom(clazz))) return true;
 
         return false;
     }
@@ -547,7 +660,7 @@ public abstract class StorageContainer {
         }, (container, recompose) -> {
             recompose.addRecomposeKey("value", enumType, container::retrieveAsync);
 
-            return recompose.onComplete(completed -> Enum.valueOf(enumType, (String) completed.getCompletedObject("value")));
+            return recompose.onComplete(completed -> Enum.valueOf(enumType, completed.getCompletedObject("value")));
         }));
     }
 
@@ -573,7 +686,68 @@ public abstract class StorageContainer {
         return storageDecomposers;
     }
 
-    public static class Recursive<T> {
-        public T function;
+    /**
+     * Stores data, but does not call {@link StoreHook#complete()} when finished.
+     *
+     * @param path The path to store the data to.
+     * @param data The data to store in the file format.
+     */
+    private CompletableFuture<Void> storeData(String path, Object data) {
+        Objects.requireNonNull(path, "Cannot store to null path!");
+
+        CompletableFuture<Void> future = new CompletableFuture<>();
+
+        if (data != null && !isSupported(data.getClass())) {
+            throw new IllegalArgumentException(data.getClass() + " is not a supported storage class. Create a StorageDecomposer to implement behavior!");
+        }
+
+        StorageDecomposer<?> decomposer = data == null ? null : getDecomposer(data.getClass());
+        DecomposedObject object = null;
+
+        //Cache the data.
+        if (decomposer != null) {
+            String decomposePath = path + ".%path%";
+            object = decomposer.applyDecompose(data);
+            for (String decomposedPath : object.getDecomposedMap().keySet()) {
+                this.cacheStore(decomposePath.replace("%path%", decomposedPath), object.getDecomposedMap().get(decomposedPath));
+            }
+        } else {
+            this.cacheStore(path, data);
+        }
+
+        DecomposedObject finalObject = object;
+        Bukkit.getScheduler().runTaskAsynchronously(StorageAPI.getPlugin(), () -> {
+            if (decomposer != null) {
+                //Add the placeholder for each decomposed path.
+                String decomposePath = path + ".%path%";
+                for (String decomposedPath : finalObject.getDecomposedMap().keySet()) {
+                    String compiledPath = decomposePath.replace("%path%", decomposedPath);
+                    Object decomposedObject = finalObject.getDecomposedMap().get(decomposedPath);
+
+                    if(decomposedObject != null) {
+                        if(this instanceof ArrayContainer<?> arrayContainer && arrayContainer.isList(decomposedObject)) {
+                            this.storeHook().getConsumer().accept(compiledPath, arrayContainer.handleList(decomposedObject));
+                            continue;
+                        } else if(StorageContainer.getDecomposer(decomposedObject.getClass()) != null) {
+                            this.storeData(compiledPath, decomposedObject);
+                            continue;
+                        }
+                        this.storeHook().getConsumer().accept(compiledPath, decomposedObject);
+                    }
+                }
+            } else if (this instanceof ArrayContainer<?> arrayContainer && arrayContainer.isList(data)) {
+                this.storeHook().getConsumer().accept(path, arrayContainer.handleList(data));
+            } else {
+                this.storeHook().getConsumer().accept(path, data);
+            }
+
+            future.complete(null);
+        });
+
+        //No need to update these later, since this should overwrite them.
+        this.getCache().getSetCache().remove(path);
+        this.getCache().getRemoveCache().remove(path);
+
+        return future;
     }
 }
